maybe fix leak

+++ src/Models/Model_Tests.cpp
@@ -162,7 +162,7 @@ TEST(MODEL_BASE, Prediction){
     0.3525;
     MatrixXf p = m.predict();
     TEST_MATRIX_NEAR(p, y, 1e-3);
-    MatrixXd q = m.predict(X);
+    MatrixXf q = m.predict(X);
     TEST_MATRIX_NEAR(q, y, 1e-3);
 }
